in pci_device_netbsd_probe() probe the expansion ROM and size. if it's there, set priv->rom_base and device->rom_size.
originally from mhitch, and it's basically the same code as present in
both freebsd_pci.c and openbsd_pci.c.

add a comment about needing to avoid the default bios mapping if this
card is not the console.

mmap() with MAP_SHARED, since the kernel forces this anyway, and it is
again the same in both freebsd_pci.c and openbsd_pci.c.